requirements.txt
aiogram==2.25.1
tweepy==4.14.0
vk_api==11.9.9
python-dotenv==0.21.1
schedule==1.2.0
aiohttp==3.8.5  
